requirements.txt
# =============================================================================
# ShopUNow AI Assistant - Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core
# -----------------------------------------------------------------------------
pydantic>=2.5.0
pydantic-settings>=2.1.0

# -----------------------------------------------------------------------------
# LLM Providers
# -----------------------------------------------------------------------------
openai>=1.12.0
google-genai>=1.0.0
groq>=0.4.2

# -----------------------------------------------------------------------------
# Embeddings
# -----------------------------------------------------------------------------
sentence-transformers>=2.3.0
cohere>=4.47

# -----------------------------------------------------------------------------
# Vector Store
# -----------------------------------------------------------------------------
chromadb>=0.4.22

# -----------------------------------------------------------------------------
# API Framework
# -----------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Optional: Production
# -----------------------------------------------------------------------------
# gunicorn>=21.2.0        # Production WSGI server
# redis>=5.0.0            # Session storage
# prometheus-client>=0.19 # Metrics
# structlog>=24.1.0       # Structured logging
